requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
supabase==0.7.1
pytest==6.2.5
pytest-cov==2.12.1
gunicorn==20.1.0
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
MarkupSafe==2.0.1
requests==2.26.0
email-validator==1.1.3
python-jose==3.3.0
